[Refactor][BLE] Added enhanced checks to ensure only valid devices are listed
