+++ Helping Hand/Controllers/Bluetooth/Delagate/CentralManagerDelegate.swift
@@ -1,49 +1,52 @@
-/////  CoreBluetooth Manager Delegate API implementation. Handles scanning, exploring, and connecting to BLE devices as well as the data transfer protocol
-/////  used in Helping Hand
-//import CoreBluetooth
-//import os
-//
-//extension ViewController: CBCentralManagerDelegate {
-//
-//    // Mark: Manager Status Checking
-//    /// Handles views and alert actions upon the Bluetooth system updating status
-//    func centralManagerDidUpdateState(_ central: CBCentralManager) {
-//        updateUIOnBluetoothState(state: central.state)
-//    }
-//    
-//    func centralManager(_ central: CBCentralManager, didDiscover peripheral: CBPeripheral, advertisementData: [String: Any], rssi RSSI: NSNumber) {
-//        
-//        self.peripheral = peripheral
-//        if !likelyPeripheralArray.contains(self.peripheral)
-//        {
-//            os_log("Likely peripheral discovered: %@ ...Attempting connection", peripheral)
-//            likelyPeripheralArray.append(self.peripheral)
-//            centralManager.connect(self.peripheral, options: nil)
-//        }
-//    }
-//    
-//    /*
-//     *  If the connection fails for whatever reason, we need to deal with it.
-//     */
-//    func centralManager(_ central: CBCentralManager, didFailToConnect peripheral: CBPeripheral, error: Error?) {
-//        os_log("Failed to connect to %@. %s", peripheral, String(describing: error))
-//        cleanup()
-//    }
-//    
-//    func centralManager(_ central: CBCentralManager, didConnect peripheral: CBPeripheral) {
-//        os_log("Peripheral Connected")
-//        stopScanning()
-//        os_log("Scanning Stopped")
-//        
-//        peripheral.delegate = self
-//        peripheral.discoverServices(CBUUIDs.serviceUUIDs)
-//    }
-//    
-//    /*
-//     *  Once the disconnection happens, we need to clean up our local copy of the peripheral
-//     */
-//    func centralManager(_ central: CBCentralManager, didDisconnectPeripheral peripheral: CBPeripheral, error: Error?) {
-//        os_log("Perhiperal Disconnected")
-//        cleanup()
-//    }
-//}
+import CoreBluetooth
+import os
+
+// MARK: - CBCentralManagerDelegate
+extension BluetoothManager: CBCentralManagerDelegate {
+    
+    func centralManagerDidUpdateState(_ central: CBCentralManager) {
+        bluetoothState = central.state
+    }
+    
+    func centralManager(_ central: CBCentralManager, didDiscover peripheral: CBPeripheral, advertisementData: [String: Any], rssi RSSI: NSNumber) {
+        guard peripheralInfo[peripheral] == nil else { return }
+        
+        // Pre-filter: skip devices that don't advertise any services
+        guard advertisementData[CBAdvertisementDataServiceUUIDsKey] != nil else { return }
+        
+        os_log("Discovered peripheral: %@ - starting validation", peripheral)
+        startValidation(for: peripheral)
+    }
+    
+    func centralManager(_ central: CBCentralManager, didFailToConnect peripheral: CBPeripheral, error: Error?) {
+        handleConnectionError(for: peripheral, error: error)
+    }
+    
+    func centralManager(_ central: CBCentralManager, didConnect peripheral: CBPeripheral) {
+        os_log("Connected to peripheral: %@", peripheral)
+        
+        if peripheral == mainPeripheral {
+            connectionState = .connected
+            connectedPeripheral = peripheral
+            stopScanning()
+        }
+        
+        peripheral.delegate = self
+        peripheral.discoverServices(CBUUIDs.serviceUUIDs)
+    }
+    
+    func centralManager(_ central: CBCentralManager, didDisconnectPeripheral peripheral: CBPeripheral, error: Error?) {
+        os_log("Disconnected from peripheral: %@", peripheral)
+        
+        if isValidationConnection(peripheral) {
+            // Validation was interrupted - mark as invalid unless already validated
+            if peripheralInfo[peripheral]?.validationState == .validating {
+                handleValidationResult(for: peripheral, isValid: false, reason: "disconnected during validation")
+            }
+        } else if peripheral == mainPeripheral {
+            connectionState = .disconnected
+            connectedPeripheral = nil
+            cleanup()
+        }
+    }
+}
